requirements.txt
paho-mqtt
donttrust
prisma
bcrypt
pyjwt